Restructure component directory to split Pages and generic UI components
