+++ src/CrowdApp.routing.js
@@ -1,8 +1,8 @@
 /* eslint-disable react/display-name */
-import EventsPage from "./components/EventsPage/EventsPage";
-import NewEventPage from "./components/NewEventPage/NewEventPage";
+import EventsPage from "./components/Pages/EventsPage/EventsPage";
+import NewEventPage from "./components/Pages/NewEventPage/NewEventPage";
 import React from "react";
-import AddRegionsPage from "./components/AddRegionsPage/AddRegionsPage";
+import AddRegionsPage from "./components/Pages/AddRegionsPage/AddRegionsPage";
 
 const routes = [
   {
@@ -45,8 +45,9 @@ const routes = [
   {
     path: "/regions",
     name: "Regions",
+    description: "Edit your region names, type and radii.",
     iconName: "object-ungroup",
-    content: () => <h1>Regions</h1>,
+    content: AddRegionsPage,
     inSidebar: true
   },
   {
