--- src/components/AddRegionsPage/AddRegionsMap/RegionMarker/RegionMarker.module.scss
@@ -1,1 +1,0 @@
-@import "../../../../styles/_vars.scss";
