--- src/components/AddRegionsPage/AddRegionsPage.module.scss
@@ -1,5 +1,0 @@
-@import "../../styles/_vars.scss";
-
-.mapContainer {
-  height: 100%;
-}
