+++ src/components/Pages/AddRegionsPage/AddRegionsMap/AddRegionsMap.js
