+++ src/components/Pages/AddRegionsPage/AddRegionsMap/AddRegionsMap.module.scss
