+++ src/components/Pages/AddRegionsPage/AddRegionsMap/CreateRegionForm/CreateRegionForm.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/_vars.scss";
+@import "../../../../../styles/_vars.scss";
 
 @mixin labelStyle {
   font: {
