+++ src/components/Pages/AddRegionsPage/AddRegionsMap/CreateRegionPopup/CreateRegionPopup.module.scss
@@ -1,0 +1,1 @@
+@import "../../../../../styles/_vars.scss";
