+++ src/components/Pages/AddRegionsPage/AddRegionsMap/RegionMarker/RegionMarker.module.scss
@@ -1,0 +1,1 @@
+@import "../../../../../styles/_vars.scss";
