+++ src/components/Pages/AddRegionsPage/AddRegionsPage.module.scss
@@ -1,1 +1,5 @@
 @import "../../../styles/_vars.scss";
+
+.mapContainer {
+  height: 100%;
+}
