+++ src/components/Pages/EventsPage/EventsPage.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/_vars.scss";
+@import "../../../styles/_vars.scss";
 
 .pageTitle {
   display: inline-block;
