+++ src/components/Pages/NewEventPage/NewEventForm.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/_vars.scss";
+@import "../../../styles/_vars.scss";
 
 .formGroups {
   margin-top: 30px;
