+++ src/components/UI/Logo/logo.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/_vars.scss";
+@import "../../../styles/_vars.scss";
 
 .logoContainer {
   color: $highlight;
