+++ src/components/UI/Page/Page.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/_vars.scss";
+@import "../../../styles/_vars.scss";
 
 .flexContainer {
   display: flex;
